Migrate popup behaviors to awe-ui
@sambaker, @ericjuvetartefact, please have a look at this. I'm
implementing pop-up as a UI behavior that we can attach to an HTML
element that is defined externally. This means that you are free to
define, style, position the element as you wish and you wrap it with
Awe.uiPopup behavior as needed. I still need to look at how this
extends to more complex scenarios, like say dynamic positioning,
dynamic borders with nipples, etc. so this is just a starting point.
Early comments are welcome and requested.