Add basic tracing to Awe.StateMachine
To enable tracing,

    yourStateMachine.tracing = true;

Tracing is turned off by default. At the moment. We aren't tracing the
requestTransition callbacks yet. @sambaker, can you have a look at
adding that? I'm not using that feat use in benchpress at the moment
and don't have a good way of testing it out.

I've also updated some of the argument validation logic and made it
throw instead of return false for scenarios that should not be treated
as runtime errors. For example, if you request a state that is not in
the array of states, we now throw instead of treating it like a
disallowed transition.