Ensure consistent drag position coordinates
After a discussion with @sambaker about the possible inconsistencies
here, I decided to change this to always return pageX and pageY
coordinates for the drag events.

If a specific implementation needs to map these to something else, they
can implement a drag filter that transforms the values to fit their
needs.

@sambaker, please review.